fix failure to compile pict import plugin with msvc caused by ambiguity with QByteArray::operator[]
git-svn-id: svn://scribus.net/trunk/Scribus@14006 11d20701-8431-0410-a711-e3c959e3b870

+++ scribus/plugins/pctimplugin/importpct.cpp
@@ -1592,7 +1592,7 @@ void PctPlug::handlePixmap(QDataStream &ts, quint16 opCode)
 					if (component_size == 4)
 					{
 						uchar *q = (uchar*)(image.scanLine(rr));
-						for (quint16 xx = 0; xx < img.size(); xx++)
+						for (int xx = 0; xx < img.size(); xx++)
 						{
 							uchar i = (img[xx] >> 4) & 0x0F;
 							uchar j = img[xx] & 0x0F;
@@ -1623,7 +1623,7 @@ void PctPlug::handlePixmap(QDataStream &ts, quint16 opCode)
 				else if (component_size == 8)
 				{
 					QRgb *q = (QRgb*)(image.scanLine(rr));
-					for (quint16 xx = 0; xx < pixCols; xx++)
+					for (uint xx = 0; xx < (uint) pixCols; xx++)
 					{
 						uchar r, g, b;
 						uchar a = 255;
